Fix for Python 3.7
Removed uses of walrus operator and @functools.cached_property.

Annoyingly there is (I think) a bug in using tox and Poetry
together (https://github.com/python-poetry/poetry/issues/6970)
where when following instructions as described in Poetry FAQ
(sanity check!) Poetry ignores the Python version of the tox
environment.

Because of this I didn't notice 3.7 was bad until it was
caught in CI (I copied much of the code in from other project of
mine).